chore: add logging types, make log writer injectable and generic, use mongo log writer only at binary
