+++ packages/core/package.json
@@ -27,11 +27,11 @@
   },
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.29.0",
-    "@mongodb-js/mcp-types": "workspace:*",
     "mongodb-redact": "^1.1.6"
   },
   "devDependencies": {
     "@mongodb-js/mcp-logging": "workspace:*",
+    "@mongodb-js/mcp-types": "workspace:*",
     "@types/node": "^25.6.0",
     "typescript": "^5.9.2",
     "vitest": "^4.1.4"
