+++ packages/core/src/index.ts
@@ -1,6 +1,4 @@
 export { LoggerBase } from "./logging/loggerBase.js";
-export { McpLogger } from "./logging/mcpLogger.js";
-export { ConsoleLogger } from "./logging/consoleLogger.js";
 export { NoopLogger } from "./logging/noopLogger.js";
 export { CompositeLogger } from "./logging/compositeLogger.js";
 export { Keychain, registerGlobalSecretToRedact } from "./keychain.js";
