+++ packages/logging/src/index.ts
@@ -1,4 +1,5 @@
 export { DiskLogger } from "./diskLogger.js";
+export type { DiskLoggerOptions } from "./diskLogger.js";
 export { LogId } from "./logId.js";
-export type { MongoLogId } from "mongodb-log-writer";
-export type { MongoLogPayload } from "./types.js";
+export { McpLogger } from "./mcpLogger.js";
+export { ConsoleLogger } from "./consoleLogger.js";
