--- packages/logging/src/types.ts
@@ -1,4 +1,0 @@
-import type { MongoLogId } from "mongodb-log-writer";
-import type { LogPayload } from "@mongodb-js/mcp-types";
-
-export type MongoLogPayload = LogPayload<MongoLogId>;
