+++ packages/types/package.json
@@ -10,7 +10,6 @@
       "default": "./dist/index.js"
     }
   },
-  "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "files": [
     "dist"
