+++ src/common/exportsManager.ts
@@ -10,7 +10,8 @@ import { Transform } from "stream";
 import { pipeline } from "stream/promises";
 import type { UserConfig } from "./config/userConfig.js";
 import { type LoggerBase } from "@mongodb-js/mcp-core";
-import { LogId, type MongoLogId } from "@mongodb-js/mcp-logging";
+import { LogId } from "@mongodb-js/mcp-logging";
+import type { MongoLogId } from "@mongodb-js/mcp-types";
 
 export const jsonExportFormat = z.enum(["relaxed", "canonical"]);
 export type JSONExportFormat = z.infer<typeof jsonExportFormat>;
